+++ _codegen/cog_sources/ChunksmithForge.java
@@ -216,7 +216,8 @@ public final class ChunksmithForge {
         registerArguments(command, literal(CommandLiteral.CENTER),
                 argument(CommandLiteral.X, word()),
                 argument(CommandLiteral.Z, word()));
-        registerArguments(command, literal(CommandLiteral.DEBUG));
+        registerArguments(command, literal(CommandLiteral.DEBUG),
+                argument(CommandLiteral.VALUE, word()));
         registerArguments(command, literal(CommandLiteral.CONFIRM));
         registerArguments(command, literal(CommandLiteral.CONTINUE),
                 argument(CommandLiteral.WORLD, dimension()));