+++ _codegen/cog_sources/ChunksmithNeoForge.java
@@ -178,7 +178,8 @@ public class ChunksmithNeoForge {
         registerArguments(command, literal(CommandLiteral.CENTER),
                 argument(CommandLiteral.X, word()),
                 argument(CommandLiteral.Z, word()));
-        registerArguments(command, literal(CommandLiteral.DEBUG));
+        registerArguments(command, literal(CommandLiteral.DEBUG),
+                argument(CommandLiteral.VALUE, word()));
         registerArguments(command, literal(CommandLiteral.CONFIRM));
         registerArguments(command, literal(CommandLiteral.CONTINUE),
                 argument(CommandLiteral.WORLD, dimension()));